feat: Enable scheduled task for making HTTP requests to Ethoca
This commit enables the scheduled task for making HTTP requests to Ethoca by uncommenting the necessary code and updating the command to `app:make-http-request`. The task is set to run every five seconds and the output is appended to the `test.log` file.

This change is made to support the integration of Ethoca alerts into the application.