Refactor EthocaAlert creation in MakeSoapAlertsRequest
The code changes in `MakeSoapAlertsRequest.php` refactor the way EthocaAlert objects are created and stored in the database. Instead of directly creating the EthocaAlert object and saving it, a new method `mapAlertResponse` is added to the EthocaAlert model to handle the mapping of the alert response fields. The mapped EthocaAlert object is then saved to the database.