Refactor ethoca_requests migration to add default values and make search dates nullable
The code changes involve refactoring the ethoca_requests migration. Default values are added for the 'alert_type' column, setting it to 'all'. The 'search_start_date' and 'search_end_date' columns are made nullable and given default values of null. These changes aim to improve the flexibility and usability of the ethoca_requests table.