feat: Add 'code' column to 'crm_actions' table
This commit adds a new 'code' column to the 'crm_actions' table in the database/migrations/2024_05_14_141942_create_crm_actions_table.php file. The 'code' column is a nullable tiny integer that represents the CRM action code for this app. This addition enhances the functionality of the CRM system by allowing for more specific categorization of actions.