feat: Add constant for KK API URL in ProcessAlert job
This commit adds a constant `KK_API_URL` to the `ProcessAlert` job in the `app/Jobs/ProcessAlert.php` file. The constant represents the URL for the KK API and is set to `'https://api.konnektive.com/'`. This addition improves the readability and maintainability of the code by centralizing the API URL and making it easily configurable.